Move core scanner and browser into libraries.
