+++ apps/api/src/results/results.controller.spec.ts
@@ -33,13 +33,6 @@ describe('ResultsController', () => {
 
   it('should return a list of results', async () => {
     const coreResult = new CoreResult();
-    const date = new Date();
-    coreResult.agency = 'GSA';
-    coreResult.branch = 'Executive';
-    coreResult.created = date.toDateString();
-    coreResult.updated = date.toDateString();
-    coreResult.id = 1;
-    coreResult.targetUrl = 'https://18f.gov';
     coreResult.finalUrl = 'https://18f.gsa.gov';
     const expected = [coreResult];
     mockResultsService.findAll.calledWith().mockResolvedValue(expected);
