+++ apps/scan-engine/src/scan-engine.consumer.spec.ts
@@ -1,16 +1,16 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CoreScanner } from '../scanners/core/core.scanner';
-import { CoreInputDto } from '../../../../dtos/scanners/core.input.dto';
-import { CoreOutputDto } from '../../../../dtos/scanners/core.output.dto';
 import { mock, mockReset, MockProxy } from 'jest-mock-extended';
-import { Scanner } from '../scanners/scanner.interface';
 import { ScanEngineConsumer } from './scan-engine.consumer';
 import { Job } from 'bull';
 import { CoreResultService } from '@app/database/core-results/core-result.service';
 import { CreateCoreResultDto } from '@app/database/core-results/dto/create-core-result.dto';
 import { LoggerService } from '@app/logger';
+import { CoreInputDto } from 'common/dtos/scanners/core.input.dto';
+import { CoreOutputDto } from 'common/dtos/scanners/core.output.dto';
+import { Scanner } from 'common/interfaces/scanner.interface';
+import { CoreScannerService } from '@app/core-scanner';
 
-describe('ScanEngineController', () => {
+describe('ScanEngineConsumer', () => {
   let consumer: ScanEngineConsumer;
   let module: TestingModule;
   let mockCoreScanner: MockProxy<Scanner<CoreInputDto, CoreOutputDto>>;
@@ -27,7 +27,7 @@ describe('ScanEngineController', () => {
       providers: [
         ScanEngineConsumer,
         {
-          provide: CoreScanner,
+          provide: CoreScannerService,
           useValue: mockCoreScanner,
         },
         {
