--- apps/scanner/src/app.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
-import { ScanEngineModule } from './scan-engine/scan-engine.module';
-import { ScannersModule } from './scanners/scanners.module';
-
-@Module({
-  imports: [ScanEngineModule, ScannersModule, ConfigModule.forRoot()],
-  controllers: [],
-  providers: [],
-})
-export class AppModule {}
