--- apps/scanner/src/scanners/scanners.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BrowserFactoryProvider } from './browser.provider';
-import { CoreScanner } from './core/core.scanner';
-
-@Module({
-  providers: [BrowserFactoryProvider, CoreScanner],
-  exports: [CoreScanner, BrowserFactoryProvider],
-})
-export class ScannersModule {}
