+++ common/dtos/scanners/core.input.dto.ts
