+++ common/dtos/scanners/core.output.dto.ts
