+++ common/interfaces/scanner.interface.ts
