+++ libs/browser/src/browser.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { BrowserService } from './browser.service';
+
+@Module({
+  providers: [BrowserService],
+  exports: [BrowserService],
+})
+export class BrowserModule {}
