+++ libs/browser/src/index.ts
@@ -1,0 +1,2 @@
+export * from './browser.module';
+export * from './browser.service';
