+++ libs/core-scanner/src/index.ts
@@ -1,0 +1,2 @@
+export * from './core-scanner.module';
+export * from './core-scanner.service';
